+++ back/src/util/path.go
@@ -4,6 +4,6 @@ import "path/filepath"
 
 // 戻り値の例: "/public/uploads/poster/myfile.png"
 // fileがmyfile.pngになる
-func BuildPosterURL(filename string) string {
+func BuildPosterPath(filename string) string {
 	return filepath.ToSlash(filepath.Join("/public/uploads/posters", filename))
 }
