+++ front/src/app/tickets/schedule/components/MovieTL.tsx
